Fix more imports for extract.utils
