+++ src/textacy/extract/keyterms/graph_base.py
@@ -8,7 +8,6 @@ from typing import (
     Any,
     Callable,
     Collection,
-    DefaultDict,
     Dict,
     Optional,
     Sequence,
@@ -22,7 +21,7 @@ from cytoolz import itertoolz
 from spacy.tokens import Span, Token
 
 from ... import errors
-from . import utils as kt_utils
+from .. import utils as ext_utils
 
 LOGGER = logging.getLogger(__name__)
 
@@ -84,7 +83,7 @@ def build_graph_from_terms(
         windows = itertoolz.sliding_window(window_size, terms)
     elif isinstance(first_term, (Span, Token)):
         windows = itertoolz.sliding_window(
-            window_size, kt_utils.normalize_terms(terms, normalize)
+            window_size, ext_utils.normalize_terms(terms, normalize)
         )
     else:
         raise TypeError(
